REFACTOR: 딥링크 로직 중복 코드 개선
